5.3 Init Database though SQL (#5)

+++ jpa-reader-writer/src/main/resources/jdbc/products-schema.sql
@@ -1,9 +1,9 @@
 create table if not exists products_jpa
 (
-    prod_id   int         not null auto_increment,
-    prod_name varchar(50)  not null,
-    prod_desc varchar(100) not null,
-    price     double       not null,
-    unit      long         not null,
+    prod_id   int            not null auto_increment,
+    prod_name varchar(50)    not null,
+    prod_desc varchar(100)   not null,
+    price     decimal(19, 2) not null,
+    unit      long           not null,
     primary key (prod_id)
 );
